--- src/components/MediaPage/styles.js
@@ -1,11 +1,0 @@
-import { makeStyles } from '@material-ui/core/styles';
-
-const useStyles = makeStyles({
-  holder: {
-    display: 'flex',
-    flexFlow: 'row wrap',
-    maxWidth: '1200px',
-  },
-});
-
-export default useStyles;
